requirements.txt
pymodbus==3.6.8
pyodbc
PyYAML
python-dotenv
pytest
ruff